Refactor installation system for improved organization
This commit restructures the installation system of the Windows Screen Logger application. Key changes include the introduction of a new namespace `WindowsScreenLogger.Installation`, movement of relevant files to a dedicated `Installation` folder, and embedding installation scripts as resources. Classes such as `SelfInstaller`, `WindowsAppsRegistry`, and `StartupRegistry` have been updated for better modularity and clarity. The installation process now runs without requiring administrator rights, enhancing user experience. Additionally, comprehensive documentation has been added, and error handling has been improved for better reliability during installation and uninstallation.
